Extract shared goal measurement helpers
